workshops/ec2-auto-scaling-with-multiple-instance-types-and-purchase-options/cdk/requirements.txt
aws-cdk.core

aws-cdk.aws-ec2
aws-cdk.aws-cloud9
aws-cdk.aws-iam
